ai_backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
ultralytics
opencv-python
Pillow
numpy
torch
torchvision
segment-anything
scikit-learn
webcolors
requests
python-dotenv
gunicorn